Adds an ELN schema that allows to create entries that clone data from a labfolder project.
Simple ELN entries can be created that act as an interface with labfolder project. The ELN schema allows to add a labfolder URL and user credentials. The processing will uses this information and the labfolder API to populate the entry with a copy of all data from this project, including downloading images and data-files.

Changelog: feature