Changed gui artifact generation to happen at app start.
This removes the gui artifacts from the gui build. The artifacts are loaded from the application and therefore depend on the application configuration and not the current build. This should allow us to load plugins at application start that influece the gui artifacts.
Include prod gui build source maps.

Changelog: Changed