Dialog to create entry
A dialog designed to select schema from the built-in schemas or custom schemas.

The sectionSelectAutocomplete has been used for searching custom schemas

Tests for this dialog were added.

Many bug fixes on sectionSelectAutocomplete which is shared with reference edit quantity. The tests for the reference edit quantity have been modified based on the bug fixes.

Changelog: Added