Added Activity ID to ExperiementStep
Added a `lab_id` property to the `ExperimentStep` section for specifying the activity by readable ID.

Added a normalizer to the `ExperimentStep` section which fetches the activity reference from the readable ID and vice versa. The normlizer also fills the start time of the `ExperimentStep` with the `datetime` property of the referenced activity if no start time is given.

Changelog: Added