auth_dict, userN_auth -> auth_headers, .._tokens
Fixture auth_dict returned both a header and a token, the latter was
discarded in most tests. The specific userN_auth fixtures only returned
the header. To simplify, those fixtures are replaced with just
auth_headers for most cases, and upload_tokens for the few cases where
the token is needed.