Move visible-for-all checkbox to upload overview
The checkbox to make the update "publicly visible" was moved from the EditMembersDialog to the UploadOverview as a new step. An additional button for the dialog was added there as well.

Checking the box will reset and disable the embargo, and publishing with embargo will disable the checkbox. There were problems if both are applied, and it is an illogical state.

A new icon (red eye) shows that an upload is unpublished but "publicly visible" (published uploads without embargo show a blue globe).

Changelog: Changed